Remove magit due to inability to work with its maintainer
magit uses a package (with the same maintainer) called with-editor to
enable using emacsclient as EDITOR.  Why it needs to do this I do not
understand, since emacsclient works just fine when configured in $EDITOR
everywhere else.

This package causes C-x C-c - normally (save-buffers-kill-terminal) - to:

- Not prompt for saving the buffer contents
- Close the buffer and the pane
- Return failure to the invoking process (git)

So it throws away your data with no easy recovery, and binds this
behavior to the key normally used for safely exiting!

Upon conversation with the maintainer in
https://github.com/magit/with-editor/issues/66 they believe it will no
longer exit on C-x C-c, which is better, but still wrong because now we
have to learn an entirely new exit key.  Also: patches aren't downstream
anywhere and I'm tired of waiting for a change that won't even fix the
problem.  Magit is nice, but not *that* nice.